// client/src/shared/api/model/articleEntity.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Temp title
 * OpenAPI spec version: 1.0
 */
import type { ArticleAuthor } from './articleAuthor';
import type { ArticleEntityAuthorId } from './articleEntityAuthorId';
import type { ArticleEntityDescription } from './articleEntityDescription';

export interface ArticleEntity {
  id: number;
  title: string;
  /** @nullable */
  description?: ArticleEntityDescription;
  body: string;
  published: boolean;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  authorId?: ArticleEntityAuthorId;
  author?: ArticleAuthor | null;
}